requirements.txt
python-dotenv==0.12.0
requests==2.23.0
terminaltables==3.1.0
